If the article does not exist at the time of deletion, redirect
why
Check whether or not the article exists with `$post->exists()`, and if it is `true`, the process continues. If it is true, the process continues. If it has already been removed and the article does not exist, you can have the `root` redirect it. If you do not do it, you should be careful because an error occurs.